netdev-offload-dpdk: Refactor flow patterns and actions code
Refactor the flow patterns and actions code to a new source file for
better readability and towards adding more code to it.

Issue: 1955511
